src/auth/requirements.txt
astroid==2.15.4
click==8.1.3
dill==0.3.6
Flask==2.2.3
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector-python==8.0.33
parso==0.8.3
platformdirs==3.3.0
protobuf==3.20.3
PyJWT==2.6.0
pylint==2.17.3
PyMySQL==1.0.3
tomlkit==0.11.7
Werkzeug==2.2.3
wrapt==1.15.0
